Stop memory leak in NDArrayPool::alloc
When the caller supply a new memory buffer in the pData argument,
the old pArray->pData pointer need to be freed or it will leak as the
new pointer simply overwrite the old one, losing the only reference
to the original buffer.
We can safely free the old pArray->pData buffer as the pArray is
either a brand new one with pData==NULL or one off the NDArrayPool
free-list.