User callback when done with an NDArray
This new feature allow a user of an NDArrayPool to register a callback
to be called when an NDArray (with a user-supplied memory buffer in pData)
is no longer in use and about to be put on the free-list.
Note that the callback implementation must never block!

This is useful to let users of the NDArrayPool be notified when a bit of memory
which they supplied is no longer in use. The memory block may then be freed or
re-used.